porthole_alert/requirements.txt
# Core dependencies
requests
python-dotenv

# LangChain related
langchain
langchain-openai

# OpenAI
openai

streamlit